+++ src/angular-cesium-widgets/components.ts
@@ -1,1 +1,3 @@
 export * from './directives/draggable-to-map.directive';
+export * from './components/toolbar/ac-toolbar/ac-toolbar.component';
+export * from './components/toolbar/ac-toolbar-button/ac-toolbar-button.component';