+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Next
 ## Feat
 * Added `calcEllipseContainingRect` static function to `CesiumHeatMapMaterialCreator`
+* Added `ac-toolbar` and `ac-toolbar-button` components
+* Added toolbar example and flyTo home toolbar option
+
 
 ## 0.0.34
 ### Fixes